Move u32 cast into flag_type! macro

+++ src/sdl2/flag.rs
@@ -100,7 +100,7 @@ macro_rules! flag_type(
         }
 
         $(
-            pub static $name: $typename = $typename { bits: $value };
+            pub static $name: $typename = $typename { bits: $value as u32 };
         )+
     }
 )
